Move git subprocess env construction out of agent domain layer
reviewDiff.ts (a host-agnostic `src/agent/` module) read `process.env`
directly to strip helper-invoking keys and extend PATH for its git
subprocess. That env/PATH construction is an OS-infrastructure concern
and its sibling helper `extendEnvPath` already lives in the utils layer.

Relocate `makeMachineGitEnv` (and its unsafe-key set) to
`@utils/system/platformPaths`, co-located with `extendEnvPath`, so the
`process.env` read no longer lives in the agent domain. Behavior is
byte-identical; reviewDiff now imports the helper.

Also prefix the two bare `assert` imports in test-kernel with `node:`
to match the repo-wide node: builtin convention.
